Move types into seperate file
