+++ app/src/lib/branch/BranchLane.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
 	import BranchCard from './BranchCard.svelte';
-	import { Project } from '$lib/backend/projects';
 	import { projectLaneCollapsed } from '$lib/config/config';
 	import FileCard from '$lib/file/FileCard.svelte';
 	import { persisted } from '$lib/persisted/persisted';
+	import { Project } from '$lib/projects/types';
 	import { SETTINGS, type Settings } from '$lib/settings/userSettings';
 	import Resizer from '$lib/shared/Resizer.svelte';
 	import { getContext, getContextStoreBySymbol, createContextStore } from '$lib/utils/context';
