+++ app/src/lib/commit/CommitList.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
 	import CommitCard from './CommitCard.svelte';
-	import { Project } from '$lib/backend/projects';
 	import { transformAnyCommit } from '$lib/commitLines/transformers';
 	import InsertEmptyCommitAction from '$lib/components/InsertEmptyCommitAction.svelte';
 	import {
@@ -9,6 +8,7 @@
 	} from '$lib/dragging/reorderDropzoneManager';
 	import Dropzone from '$lib/dropzone/Dropzone.svelte';
 	import LineOverlay from '$lib/dropzone/LineOverlay.svelte';
+	import { Project } from '$lib/projects/types';
 	import { getContext } from '$lib/utils/context';
 	import { getContextStore } from '$lib/utils/context';
 	import { BranchController } from '$lib/vbranches/branchController';
