+++ app/src/lib/components/NotOnGitButlerBranch.svelte
@@ -5,8 +5,9 @@
 	import Link from '../shared/Link.svelte';
 	import ProjectNameLabel from '../shared/ProjectNameLabel.svelte';
 	import derectionDoubtSvg from '$lib/assets/illustrations/direction-doubt.svg?raw';
-	import { ProjectListingService, Project } from '$lib/backend/projects';
 	import { showError } from '$lib/notifications/toasts';
+	import { ProjectListingService } from '$lib/projects/projectListingService';
+	import { Project } from '$lib/projects/types';
 	import Button from '$lib/shared/Button.svelte';
 	import { getContext } from '$lib/utils/context';
 	import * as toasts from '$lib/utils/toasts';
@@ -36,7 +37,7 @@
 				showError('Failed to delete project', err);
 			} finally {
 				isDeleting = false;
-				projectService.reload();
+				projectService.reloadAll();
 			}
 		}
 	}
