+++ app/src/lib/components/ProblemLoadingRepo.svelte
@@ -3,8 +3,9 @@
 	import ProjectSwitcher from './ProjectSwitcher.svelte';
 	import RemoveProjectButton from './RemoveProjectButton.svelte';
 	import loadErrorSvg from '$lib/assets/illustrations/load-error.svg?raw';
-	import { ProjectListingService, Project } from '$lib/backend/projects';
 	import { showError } from '$lib/notifications/toasts';
+	import { ProjectListingService } from '$lib/projects/projectListingService';
+	import { Project } from '$lib/projects/types';
 	import Icon from '$lib/shared/Icon.svelte';
 	import { getContext } from '$lib/utils/context';
 	import * as toasts from '$lib/utils/toasts';
@@ -30,7 +31,7 @@
 			showError('Failed to delete project', err);
 		} finally {
 			loading = false;
-			projectService.reload();
+			projectService.reloadAll();
 		}
 	}
 </script>
