+++ app/src/routes/+page.svelte
@@ -1,11 +1,11 @@
 <script lang="ts">
 	import analyticsSvg from '$lib/assets/illustrations/analytics.svg?raw';
 	import newProjectSvg from '$lib/assets/illustrations/new-project.svg?raw';
-	import { ProjectListingService } from '$lib/backend/projects';
 	import DecorativeSplitView from '$lib/components/DecorativeSplitView.svelte';
 	import FullviewLoading from '$lib/components/FullviewLoading.svelte';
 	import Welcome from '$lib/components/Welcome.svelte';
 	import { appAnalyticsConfirmed } from '$lib/config/appSettings';
+	import { ProjectListingService } from '$lib/projects/projectListingService';
 	import AnalyticsConfirmation from '$lib/settings/AnalyticsConfirmation.svelte';
 	import { getContext } from '$lib/utils/context';
 	import { derived } from 'svelte/store';
